requirements-test.txt
#https://playwright.dev/python/docs/intro#updating-playwright
-r requirements.txt
pytest-playwright
pytest